Use drive root rather than game folder as default Project_127_Files path recommendation on XboxPC
Reinstalling the game on XboxPC results in duplicate folders being created by the Xbox app as Project_127_Files folder is not deleted during uninstallation, and while reinstalling it will create a new one rather than using that. 

To avoid this, default Project_127_Files path recommendation is changed to drive root for XboxPC users.